# Strip comments in the generated editor/documentation translation header