This reduces the total generated header size from 35 MB to 28.7 MB (-6.3 MB).

Make sure to `rm -f editor/*translations.gen.h` then recompile to see a change in binary size locally.

This partially addresses https://github.com/godotengine/godot-proposals/issues/3421.

### Before

```md
17,457,698 hugo  editor/doc_translations.gen.h
17,544,165 hugo  editor/editor_translations.gen.h
```

### After

```md
15,107,211  editor/doc_translations.gen.h
13,613,454  editor/editor_translations.gen.h
```