@akien-mga 2021-11-01
Did you see what difference it makes on the final build? And on build speed? I thought about doing this but was afraid that a line-by-line `for` loop would really slow things down significantly.