@Calinou 2021-11-02
File size difference (after rebasing this PR on `master` and `strip`ping both binaries):

```
153,532,032 hugo  2 Nov 20:20  godot.linuxbsd.tools.64.llvm (this PR)
154,639,680 hugo  2 Nov 20:20  godot.linuxbsd.tools.64.llvm.vanilla (`master`)
```

### With this PR rebased on `master`

```
Benchmark #1: pyston-scons -j10 platform=linuxbsd use_llvm=yes use_lld=yes werror=no tests=no
  Time (mean ± σ):     40.267 s ±  1.726 s    [User: 40.361 s, System: 1.339 s]
  Range (min … max):   38.470 s … 44.771 s    10 runs
```

## `master` branch

```
Benchmark #1: pyston-scons -j10 platform=linuxbsd use_llvm=yes use_lld=yes werror=no tests=no
  Time (mean ± σ):     14.192 s ±  1.291 s    [User: 14.473 s, System: 1.151 s]
  Range (min … max):   13.435 s … 17.754 s    10 runs
```

This PR does slow down initial build times quite a bit. I'm not sure how we can improve this.