@akien-mga 2021-11-02
Yeah that slowdown is not worth the small size gain. It would need to be a fast file operation, maybe regex or grepping, and not line by line byte operations.