@akien-mga 2021-12-02
Closing as the impact on build time is not worth it for that size gain.